Assignment 2

+++ events/views.py
@@ -2,18 +2,28 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.http import HttpResponse
 from events import models, forms
 from django.contrib import messages
+from django.contrib.auth.models import User, Permission, Group
 from django.db.models import Prefetch, Count, Q
 from datetime import datetime
+from django.conf import settings
+from django.core.mail import send_mail
+from django.contrib.auth.decorators import login_required, user_passes_test, permission_required
+from users.views import is_admin
 
+def is_organizer(user):
+    return user.groups.filter(name="Organizer").exists()
+
+def is_participant(user):
+    return user.groups.filter(name="Participant").exists()
 
 # Create your views here.
 def event_list(request):
     # events = models.Event.objects.all()
     events = models.Event.objects.select_related("category").annotate(
-        Count("participant")
+        Count("participants")
     )
     data = "abc"
-    if request.method == "POST":
+    if request.method == "POST": 
         data = request.POST
         if data["name"] != "" and data["category"] != "all":
             # category = models.Category.get(name=)
@@ -28,17 +38,38 @@ def event_list(request):
     context = {"events": events, "search_form": search_form, "data": data}
     return render(request, "event/events-list.html", context)
 
+def event_cards(request):
+    # events = models.Event.objects.all()
+    events = models.Event.objects.select_related("category").annotate(
+        Count("participants")
+    )
+    data = "abc"
+    if request.method == "POST": 
+        data = request.POST
+        if data["name"] != "" and data["category"] != "all":
+            # category = models.Category.get(name=)
+            events = events.filter(
+                name__icontains=data["name"], category__name=data["category"]
+            )
+        elif data["name"] != "":
+            events = events.filter(name__icontains=data["name"])
+        elif data["category"] != "all":
+            events = events.filter(category__name=data["category"])
+    context = {"events": events}
+    return render(request, "event/event-cards.html", context)
+
 
 def event_details(request, pk):
     event = models.Event.objects.prefetch_related("participant_set").get(id=pk)
     return render(request, "events/event_detail.html", {"event": event})
 
-
+@login_required
+@permission_required("events.add_event", login_url="no-permission")
 def event_create(request):
     event_form = forms.EventForm()
 
     if request.method == "POST":
-        event_form = forms.EventForm(request.POST)
+        event_form = forms.EventForm(request.POST, request.FILES)
 
         if event_form.is_valid():
             event_form.save()
@@ -50,7 +81,8 @@ def event_create(request):
     }
     return render(request, "event/events-forms.html", context)
 
-
+@login_required
+@permission_required("events.change_event", login_url="no-permission")
 def event_update(request, pk):
     event = get_object_or_404(models.Event, id=pk)
     event_form = forms.EventForm(instance=event)
@@ -66,70 +98,58 @@ def event_update(request, pk):
     context = {"event_form": event_form}
     return render(request, "event/events-forms.html", context)
 
-
+@login_required
+@permission_required("events.delete_event", login_url="no-permission")
 def event_delete(request, pk):
     event = get_object_or_404(models.Event, id=pk)
     if request.method == "POST":
         event.delete()
         messages.success(request, "Event deleted Successfully")
-        return redirect("event-list")
-    return redirect("event-list")
+        return redirect("event-cards")
+    return redirect("event-cards")
 
 
 def event_details(request, pk):
     event = get_object_or_404(models.Event, id=pk)
     return render(request, "event/event-details.html", {"event": event})
 
-
+@login_required
+@permission_required("participant.view_participant", login_url="no-permission")
 def participant_list(request):
-    participants = models.Participant.objects.prefetch_related("events").all()
+    participants = User.objects.all()
     return render(
         request, "participant/participant-list.html", {"participants": participants}
     )
 
 
-def participant_create(request):
-    participant_form = forms.ParticipantForm()
 
-    if request.method == "POST":
-        participant_form = forms.ParticipantForm(request.POST)
-
-        if participant_form.is_valid():
-            participant_form.save()
-            messages.success(request, f"Participant created Successfully")
-            return redirect("participant-create")
-
-    return render(
-        request,
-        "participant/participant-form.html",
-        {"participant_form": participant_form},
-    )
-
-
+@login_required
+@permission_required("events.change_participant", login_url="no-permission")
 def participant_update(request, pk):
-    participant = get_object_or_404(models.Participant, id=pk)
-    participant_form = forms.ParticipantForm(instance=participant)
+    events = models.Event.objects.all()
+    participant = get_object_or_404(User, id=pk)
 
     if request.method == "POST":
-        participant_form = forms.ParticipantForm(request.POST, instance=participant)
+        participant.username = request.POST.get("username")
+        # print(request.POST.getlist("events"))
+        participant.event_set.set(request.POST.getlist("events"))
+        participant.save()
+        messages.success(request, f"Participant Updated Successfully")
+        return redirect("participant-update", participant.id)
 
-        if participant_form.is_valid():
-            participant_form.save()
-            messages.success(request, f"Participant Updated Successfully")
-            return redirect("participant-create")
-
-    context = {"participant_form": participant_form}
-    return render(request, "participant/participant-form.html", context)
-
+    context = {"participant": participant, "events": events}
+    return render(request, "participant/participant-form.html", context) 
 
 def participant_delete(request, pk):
-    participant = get_object_or_404(models.Participant, id=pk)
-
-    if request.method == "POST":
-        participant.delete()
-        messages.success(request, "Participant Deleted Successfully")
-        return redirect("participant-list")
-    return redirect("participant-list")
+    try:
+        if request.method == "POST":
+            participant = User.objects.get(id=pk)
+            if participant:
+                participant.delete()
+                messages.success(request, "Deleted Successfully!!")
+    except Exception as e:
+        print(str(e))
+    return redirect("participant-list") 
 
 
 def category_list(request):
@@ -180,21 +200,21 @@ def category_delete(request, pk):
         return redirect("category-list")
     return redirect("category-list")
 
-
+@login_required
+@user_passes_test(is_organizer, login_url="no-permission")
 def organizer_dashboard(request):
     # Aggregate query to calculate the total number of participants
+    categories = models.Category.objects.prefetch_related(Prefetch("event_set")).all()
     q = request.GET.get("q", "all")
-    total_participants = models.Participant.objects.aggregate(total=Count("id"))[
-        "total"
-    ]
+    total_participants = User.objects.all().distinct().count()
     base = models.Event.objects.select_related("category")
     todays_event = base.filter(date=datetime.now().date())
     if q == "all":
-        events = base.annotate(Count("participant"))
+        events = base.annotate(Count("participants"))
     elif q == "upcoming":
-        events = base.filter(date__gt=datetime.now()).annotate(Count("participant"))
+        events = base.filter(date__gt=datetime.now()).annotate(Count("participants"))
     elif q == "past":
-        events = base.filter(date__lt=datetime.now()).annotate(Count("participant"))
+        events = base.filter(date__lt=datetime.now()).annotate(Count("participants"))
 
     counts = base.aggregate(
         total=Count("id"),
@@ -206,5 +226,95 @@ def organizer_dashboard(request):
         "events": events,
         "counts": counts,
         "todays_event": todays_event, 
+        "categories": categories,
     }
     return render(request, "dashboard/organizer-dashboard.html", context)
+
+@login_required
+def rsvp(request, id):
+    event = models.Event.objects.get(id=id)
+    event_set = request.user.event_set.filter(id=id)
+    if event_set.exists():
+        messages.warning(request, f"Event {event.name} already added!")
+        return redirect("event-list")
+    else:
+        event.participants.add(request.user)
+        messages.success(request, f"Event {event.name} added successfully!")
+        try:
+            subject = f"RSVP Confirmation for {event.name}" 
+            message = f"Hello {request.user.username},\n\nYou have successfully RSVP'd to: {event.name}\nDate: {event.date}\nTime: {event.time}\nLocation: {event.location}\n\nThank you for joining!"
+            recipient_list = [request.user.email] 
+            send_mail(subject, message, settings.EMAIL_HOST_USER, recipient_list)
+        except Exception as e:
+            print(str(e)) 
+    return redirect("event-list") 
+
+def admin_dashboard(request):
+    categories = models.Category.objects.prefetch_related(Prefetch("event_set")).all()
+    q = request.GET.get("q", "all")
+    participants = User.objects.all()
+    total_participants = User.objects.all().distinct().count()
+    base = models.Event.objects.select_related("category")
+    todays_event = base.filter(date=datetime.now().date())
+    groups = Group.objects.prefetch_related("user_set").all()
+
+    if q == "all":
+        events = base.annotate(Count("participants"))
+    elif q == "upcoming":
+        events = base.filter(date__gt=datetime.now()).annotate(Count("participants"))
+    elif q == "past":
+        events = base.filter(date__lt=datetime.now()).annotate(Count("participants"))
+
+    counts = base.aggregate(
+        total=Count("id"),
+        upcoming_events=Count("id", Q(date__gt=datetime.now())),
+        past_events=Count("id", Q(date__lt=datetime.now())),
+    )
+    context = {
+        "participants": participants,
+        "total_participants": total_participants,
+        "events": events,
+        "counts": counts,
+        "todays_event": todays_event, 
+        "categories": categories,
+        "groups": groups,
+    }
+    return render(request, "dashboard/admin-dashboard.html", context)
+
+@login_required
+def dashboard(request):
+    if is_organizer(request.user):
+        return redirect("organizer-dashboard")
+    elif is_admin(request.user):
+        return redirect("admin-dashboard")
+    else:
+        return redirect("participant-dashboard")
+
+@login_required 
+@user_passes_test(is_participant, login_url="no-permission")
+def participant_dashboard(request):
+    categories = models.Category.objects.prefetch_related(Prefetch("event_set")).all()
+    q = request.GET.get("q", "all")
+    total_participants = User.objects.all().distinct().count()
+    base = models.Event.objects.select_related("category")
+    enrolled_event = request.user.event_set.all()
+    if q == "all":
+        events = base.annotate(Count("participants"))
+    elif q == "upcoming":
+        events = base.filter(date__gt=datetime.now()).annotate(Count("participants"))
+    elif q == "past":
+        events = base.filter(date__lt=datetime.now()).annotate(Count("participants"))
+
+    counts = base.aggregate(
+        total=Count("id"),
+        upcoming_events=Count("id", Q(date__gt=datetime.now())),
+        past_events=Count("id", Q(date__lt=datetime.now())),
+    )
+    context = {
+        "total_participants": total_participants,
+        "events": events,
+        "counts": counts,
+        "enrolled_event": enrolled_event, 
+        "categories": categories,
+    }
+    return render(request, "dashboard/participant-dashboard.html", context)
